# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3
info:
  title: API de Histórico Médico
  version: 1.0.0
  description: API para gerenciar históricos médicos de pacientes.

servers:
  - url: http://localhost:8080/api

tags:
  - name: Historicos
    description: Operações relacionadas aos historicos do paciente

paths:
  /historicos:
    get:
      summary: Buscar lista de históricos
      operationId: buscarHistoricos
      tags:
        - Historicos
      parameters:
        - name: idHistorico
          in: query
          schema:
            $ref: '#/components/schemas/HistoricoId'
        - in: query
          name: idPaciente
          schema:
            $ref: '#/components/schemas/PacienteId'
      responses:
        '200':
          description: Lista de históricos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoricoResponse'
        '500':
          description: Erro interno no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Cria um novo histórico
      operationId: criarHistorico
      tags:
        - Historicos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoricoRequest'
      responses:
        '201':
          description: Histórico criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricoResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /historicos/{idHistorico}:
    put:
      summary: Atualiza um histórico
      operationId: atualizarHistorico
      tags:
        - Historicos
      parameters:
        - name: idHistorico
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/HistoricoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoricoAtualizarRequest'
      responses:
        '200':
          description: Histórico atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricoResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Histórico não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Remove um histórico
      operationId: removerHistorico
      tags:
        - Historicos
      parameters:
        - name: idHistorico
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/HistoricoId'
      responses:
        '204':
          description: Histórico removido
        '404':
          description: Histórico não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    # Identificadores
    HistoricoId:
      type: integer
      format: int64
      example: 1

    PacienteId:
      type: integer
      format: int64
      example: 1

    # Campos básicos reutilizáveis
    NomeHospital:
      type: string
      example: "Hospital Municipal"

    NomeMedico:
      type: string
      example: "Dr. Carlos Souza"

    Especialidade:
      type: string
      example: "Clínico Geral"

    Motivo:
      type: string
      example: "Febre e dor de cabeça"

    Diagnostico:
      type: string
      example: "Infecção viral"

    Prescricao:
      type: string
      example: "Dipirona 500mg a cada 8h por 5 dias"

    Observacoes:
      type: string
      example: "Paciente em bom estado geral"

    DataAtendimento:
      type: string
      format: date-time
      example: "2025-09-16T10:30:00"

    # Objeto principal
    HistoricoRequest:
      type: object
      properties:
        idHistorico:
          $ref: '#/components/schemas/HistoricoId'
        idPaciente:
          $ref: '#/components/schemas/PacienteId'
        hospital:
          $ref: '#/components/schemas/NomeHospital'
        medico:
          $ref: '#/components/schemas/NomeMedico'
        especialidade:
          $ref: '#/components/schemas/Especialidade'
        motivo:
          $ref: '#/components/schemas/Motivo'
        diagnostico:
          $ref: '#/components/schemas/Diagnostico'
        prescricao:
          $ref: '#/components/schemas/Prescricao'
        observacoes:
          $ref: '#/components/schemas/Observacoes'

    HistoricoResponse:
      type: object
      properties:
        idHistorico:
          $ref: '#/components/schemas/HistoricoId'
        idPaciente:
          $ref: '#/components/schemas/PacienteId'
        data:
          $ref: '#/components/schemas/DataAtendimento'
        hospital:
          $ref: '#/components/schemas/NomeHospital'
        medico:
          $ref: '#/components/schemas/NomeMedico'
        especialidade:
          $ref: '#/components/schemas/Especialidade'
        motivo:
          $ref: '#/components/schemas/Motivo'
        diagnostico:
          $ref: '#/components/schemas/Diagnostico'
        prescricao:
          $ref: '#/components/schemas/Prescricao'
        observacoes:
          $ref: '#/components/schemas/Observacoes'

    HistoricoAtualizarRequest:
      type: object
      properties:
        hospital:
          $ref: '#/components/schemas/NomeHospital'
        medico:
          $ref: '#/components/schemas/NomeMedico'
        especialidade:
          $ref: '#/components/schemas/Especialidade'
        motivo:
          $ref: '#/components/schemas/Motivo'
        diagnostico:
          $ref: '#/components/schemas/Diagnostico'
        prescricao:
          $ref: '#/components/schemas/Prescricao'
        observacoes:
          $ref: '#/components/schemas/Observacoes'

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-09-16T21:00:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Histórico não encontrado para o ID informado"
        path:
          type: string
          example: "/historicos/H123"